added PATCH method and actuator endpoint

+++ BugTracker/src/main/java/com/app/jira/BugTracker/service/TicketService.java
@@ -1,9 +1,13 @@
 package com.app.jira.BugTracker.service;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
+
 import com.app.jira.BugTracker.entity.Ticket;
 
 import com.app.jira.BugTracker.repository.TicketRepository;
@@ -28,22 +32,30 @@ public class TicketService {
 	}
 
 	public Ticket getTicketById(int ticketId) {
-		return ticketRepository.findById(ticketId).get();
+		if(ticketRepository.existsById(ticketId))
+		{
+			return ticketRepository.findById(ticketId).get();
+		}
+		else
+			throw new BadReqeustException("incoming id  doesn't exist");
 	}
 
 	public void addTicket(Ticket ticket) {
 		ticketRepository.save(ticket);
 	}
 
-	public void updateTicket(Integer id,Ticket ticket) {
-		if (id != ticket.getId()) {
-			throw new BadReqeustException("incoming id in body doesn't match path");
+	public void updateTicket(Integer ticketId,Ticket ticket) {
+		if (!ticketRepository.existsById(ticketId)) {
+			throw new BadReqeustException("incoming id in body doesn't exist");
 		}
 		this.ticketRepository.save(ticket);
 	}
 
 
 	public void deleteTicket(int ticketId) {
+		if (!ticketRepository.existsById(ticketId)) {
+			throw new BadReqeustException("incoming id in body doesn't exist");
+		}
 		this.ticketRepository.deleteById(ticketId);
 	}
 
@@ -62,5 +74,17 @@ public class TicketService {
 	public List<Ticket> findTicketByStatus(String status) {
 
 		return ticketRepository.findTicketByStatus(status);
+	}
+
+	public Ticket patchTicketByStatus(Integer id, String ticketStatus) {
+		Optional<Ticket> ticketOpt=ticketRepository.findById(id);
+		if(ticketOpt.isPresent())
+		{
+			Ticket ticket=ticketOpt.get();
+			ticket.setStatus(ticketStatus);
+			return ticketRepository.save(ticket);
+		}
+		else
+			throw new BadReqeustException("incoming id doesn't exist");
 	}
 }
